LPD-64261 Move logic to new component
